Split the rho clicommands.py file into individual class files and utility files. (#58)
