+++ rho/translation.py
@@ -1,0 +1,25 @@
+#
+# Copyright (c) 2009-2016 Red Hat, Inc.
+#
+# This software is licensed to you under the GNU General Public License,
+# version 2 (GPLv2). There is NO WARRANTY for this software, express or
+# implied, including the implied warranties of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
+# along with this software; if not, see
+# http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
+
+""" Rho Translation initalization utility """
+
+
+import gettext
+
+T = gettext.translation('rho', 'locale', fallback=True)
+
+
+def get_translation():
+    """Obtains the locale based translation setup for rho"""
+    if hasattr(T, 'ugettext'):
+        _ = T.ugettext
+    else:
+        _ = T.gettext
+    return _
